exercise_solutions/chapter-7/project-7.1_configure_idp/requirements.txt
requests==2.32.3
auth0-python==4.8.1
invoke==2.2.0
